@@ -1,23 +1,21 @@
 package com.example.get_hitched;
 
 import androidx.appcompat.app.AppCompatActivity;
-
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
-
 import java.util.ArrayList;
 import java.util.List;
 
 public class PlannerActivity extends AppCompatActivity {
 
-    private ListView locationListView, dressListView, menDressListView;
+    private RecyclerView locationRecyclerView, dressRecyclerView, menDressRecyclerView;
     private EditText guestCountEditText;
     private Button calculateButton, cateringButton;
     private TextView welcomeText, estimatedCostTextView;
@@ -31,15 +29,15 @@ public class PlannerActivity extends AppCompatActivity {
 
         initializeViews();
         setWelcomeMessage();
-        setupListViews();
+        setupRecyclerViews();
         setupListeners();
     }
 
     private void initializeViews() {
         welcomeText = findViewById(R.id.welcomeText);
-        locationListView = findViewById(R.id.locationListView);
-        dressListView = findViewById(R.id.dressListView);
-        menDressListView = findViewById(R.id.menDressListView);
+        locationRecyclerView = findViewById(R.id.locationRecyclerView);
+        dressRecyclerView = findViewById(R.id.dressRecyclerView);
+        menDressRecyclerView = findViewById(R.id.menDressRecyclerView);
         guestCountEditText = findViewById(R.id.editTextGuestCount);
         calculateButton = findViewById(R.id.buttonCalculate);
         cateringButton = findViewById(R.id.buttonCatering);
@@ -48,81 +46,62 @@ public class PlannerActivity extends AppCompatActivity {
 
     private void setWelcomeMessage() {
         SharedPreferences preferences = getSharedPreferences(PREFS_NAME, MODE_PRIVATE);
-        String username = ((SharedPreferences) preferences).getString("username", "User");
+        String username = preferences.getString("username", "User");
         welcomeText.setText("Welcome, " + username + "!");
     }
 
-    private void setupListViews() {
+    private void setupRecyclerViews() {
         List<SpinnerItem> locations = new ArrayList<>();
-        locations.add(new SpinnerItem("Beach ", 500));
-        locations.add(new SpinnerItem("Garden ", 300));
-        locations.add(new SpinnerItem("Hall ", 700));
-        locations.add(new SpinnerItem("Home ", 1000));
+        locations.add(new SpinnerItem("Beach", 500));
+        locations.add(new SpinnerItem("Garden", 300));
+        locations.add(new SpinnerItem("Hall", 700));
+        locations.add(new SpinnerItem("Home", 1000));
 
-        List<SpinnerItem> dresses = new ArrayList<>();
-        dresses.add(new SpinnerItem("Simple Dress ", 200));
-        dresses.add(new SpinnerItem("Elegant Dress ", 500));
-        dresses.add(new SpinnerItem("Designer Dress ", 1000));
-        dresses.add(new SpinnerItem("Custom Dress ", 800));
+        List<SpinnerItem> brideDresses = new ArrayList<>();
+        brideDresses.add(new SpinnerItem("Simple Dress", 200));
+        brideDresses.add(new SpinnerItem("Elegant Dress", 500));
+        brideDresses.add(new SpinnerItem("Designer Dress", 1000));
+        brideDresses.add(new SpinnerItem("Custom Dress", 800));
 
-        List<SpinnerItem> menDresses = new ArrayList<>();
-        menDresses.add(new SpinnerItem("Classic Suit ", 300));
-        menDresses.add(new SpinnerItem("Elegant Suit ", 600));
-        menDresses.add(new SpinnerItem("Designer Wear ", 400));
-        menDresses.add(new SpinnerItem("Custom Suit ", 200));
+        List<SpinnerItem> groomDresses = new ArrayList<>();
+        groomDresses.add(new SpinnerItem("Classic Suit", 300));
+        groomDresses.add(new SpinnerItem("Elegant Suit", 600));
+        groomDresses.add(new SpinnerItem("Designer Wear", 400));
+        groomDresses.add(new SpinnerItem("Custom Suit", 200));
 
-        CustomAdapter locationAdapter = new CustomAdapter(this, locations);
-        locationListView.setAdapter(locationAdapter);
+        locationRecyclerView.setLayoutManager(new LinearLayoutManager(this));
+        dressRecyclerView.setLayoutManager(new LinearLayoutManager(this));
+        menDressRecyclerView.setLayoutManager(new LinearLayoutManager(this));
 
-        CustomAdapter dressAdapter = new CustomAdapter(this, dresses);
-        dressListView.setAdapter(dressAdapter);
+        CustomAdapter locationAdapter = new CustomAdapter(this, locations, item -> {
+            selectedLocation = item;
+            Toast.makeText(this, "Selected Location: " + item.getName(), Toast.LENGTH_SHORT).show();
+        });
 
-        CustomAdapter menDressAdapter = new CustomAdapter(this, menDresses);
-        menDressListView.setAdapter(menDressAdapter);
+        CustomAdapter dressAdapter = new CustomAdapter(this, brideDresses, item -> {
+            selectedDress = item;
+            Toast.makeText(this, "Selected Bride Dress: " + item.getName(), Toast.LENGTH_SHORT).show();
+        });
+
+        CustomAdapter menDressAdapter = new CustomAdapter(this, groomDresses, item -> {
+            selectedMenDress = item;
+            Toast.makeText(this, "Selected Groom Dress: " + item.getName(), Toast.LENGTH_SHORT).show();
+        });
+
+        locationRecyclerView.setAdapter(locationAdapter);
+        dressRecyclerView.setAdapter(dressAdapter);
+        menDressRecyclerView.setAdapter(menDressAdapter);
     }
 
     private void setupListeners() {
-        locationListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                selectedLocation = (SpinnerItem) parent.getItemAtPosition(position);
-            }
-        });
-
-        dressListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                selectedDress = (SpinnerItem) parent.getItemAtPosition(position);
-            }
-        });
-
-        menDressListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                selectedMenDress = (SpinnerItem) parent.getItemAtPosition(position);
-            }
-        });
-
-        calculateButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                calculateCost();
-            }
-        });
-
-//        cateringButton.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                startActivity(new Intent(PlannerActivity.this, CateringActivity.class));
-//            }
-//        });
+        calculateButton.setOnClickListener(v -> calculateCost());
     }
 
     private void calculateCost() {
         String guestCountString = guestCountEditText.getText().toString();
 
         if (guestCountString.isEmpty() || selectedLocation == null || selectedDress == null || selectedMenDress == null) {
-            Toast.makeText(this, "Please fill all fields", Toast.LENGTH_SHORT).show();
+            Toast.makeText(this, "Please select all options and enter guest count", Toast.LENGTH_SHORT).show();
             return;
         }
 
@@ -135,12 +114,12 @@ public class PlannerActivity extends AppCompatActivity {
 
             double estimatedCost = calculateWeddingCost(selectedLocation, selectedDress, selectedMenDress, guestCount);
             String costDetails = "Selected Location: " + selectedLocation.getName() + "\n" +
-                    "Selected Dress: " + selectedDress.getName() + "\n" +
-                    "Selected Men's Dress: " + selectedMenDress.getName() + "\n" +
+                    "Selected Bride Dress: " + selectedDress.getName() + "\n" +
+                    "Selected Groom Dress: " + selectedMenDress.getName() + "\n" +
                     "Base Cost: $1000\n" +
                     "Location Cost: $" + selectedLocation.getCost() + "\n" +
-                    "Dress Cost: $" + selectedDress.getCost() + "\n" +
-                    "Men's Dress Cost: $" + selectedMenDress.getCost() + "\n" +
+                    "Bride Dress Cost: $" + selectedDress.getCost() + "\n" +
+                    "Groom Dress Cost: $" + selectedMenDress.getCost() + "\n" +
                     "Accommodation Cost per person: $50\n" +
                     "Total Guest Count: " + guestCount + "\n" +
                     "Estimated Cost: $" + estimatedCost;
